Upgrade package to target .net 5 and EpiServer 12
