--- src/EpiEasyEvents/Configuration.cs
@@ -1,10 +1,0 @@
-﻿namespace Forte.EpiEasyEvents
-{
-    public static class Configuration
-    {
-        /// <summary>
-        /// Dynamically switch handling all events.
-        /// </summary>
-        public static bool IsHandlingDisabled { get; set; } = false;
-    }
-}
