+++ src/EpiEasyEvents/DisabledEventsHandlerScopeConfiguration.cs
@@ -1,0 +1,7 @@
+﻿namespace Forte.EpiEasyEvents
+{
+    internal static class DisabledEventsHandlerScopeConfiguration
+    {
+        public static bool IsHandlingDisabled { get; set; } = false;
+    }
+}
