+++ src/EpiEasyEvents/EpiEasyEvents.csproj
@@ -1,44 +1,30 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
-
-    <Import Project="..\targets\AddReadmeMarkdown.targets" />
-    <Import Project="..\targets\CommonBuild.props" />
-
-    <PropertyGroup>
-        <PackageId>Forte.EpiEasyEvents</PackageId>
-        <Company>Forte_ Digital</Company>
-        <Copyright>Copyright $(CurrentYear)</Copyright>
-        <PackageDescription>Helpers classes to easily implement handlers for EpiServer events</PackageDescription>
-        <PackageVersion>1.0.0</PackageVersion>
-        
-        <PackageTags>episerver events</PackageTags>
-        
-        <PackageRequireLicenseAcceptance>false</PackageRequireLicenseAcceptance>
-        <!-- <PackageLicenseUrl></PackageLicenseUrl> -->
-        <!-- <PackageProjectUrl></PackageProjectUrl> -->
-
-        <!--Provide intellisense for developers-->
-        <GenerateDocumentationFile>true</GenerateDocumentationFile>
-
-        <!--useful to allow debugging -->
-        <IncludeSymbols>true</IncludeSymbols>
-        <IncludeSource>true</IncludeSource>
-
-        <!--do not change unless you want to deal with assembly redirects-->
-        <AssemblyVersion>1.0.0</AssemblyVersion>
-        <FileVersion>1.0.0</FileVersion>
-        <AssemblyName>Forte.EpiEasyEvents</AssemblyName>
-        <RootNamespace>Forte.EpiEasyEvents</RootNamespace>
-        <TargetFramework>net461</TargetFramework>
-    </PropertyGroup>
-
-    <ItemGroup>
-        <None Include=".template.config\template.json" />
-        <Content Include="web.config.transform" Condition="'$(EpiModuleFolderName)' != ''" pack="true" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <PackageReference Include="EPiServer.CMS" Version="11.10.6" />
-      <PackageReference Include="EPiServer.Framework" Version="11.10.6" />
-      <PackageReference Include="EPiServer.ServiceLocation.StructureMap" Version="2.0.1" />
-    </ItemGroup>
+  <Import Project="..\targets\AddReadmeMarkdown.targets" />
+  <Import Project="..\targets\CommonBuild.props" />
+  <PropertyGroup>
+    <PackageId>Forte.EpiEasyEvents</PackageId>
+    <Company>Forte_ Digital</Company>
+    <Copyright>Copyright $(CurrentYear)</Copyright>
+    <PackageDescription>Helpers classes to easily implement handlers for EpiServer events</PackageDescription>
+    <PackageVersion>2.0.0</PackageVersion>
+    <PackageTags>episerver events</PackageTags>
+    <PackageRequireLicenseAcceptance>false</PackageRequireLicenseAcceptance>
+    <!-- <PackageLicenseUrl></PackageLicenseUrl> -->
+    <!-- <PackageProjectUrl></PackageProjectUrl> -->
+    <!--Provide intellisense for developers-->
+    <GenerateDocumentationFile>true</GenerateDocumentationFile>
+    <!--useful to allow debugging -->
+    <IncludeSymbols>true</IncludeSymbols>
+    <IncludeSource>true</IncludeSource>
+    <!--do not change unless you want to deal with assembly redirects-->
+    <AssemblyVersion>1.0.0</AssemblyVersion>
+    <FileVersion>1.0.0</FileVersion>
+    <AssemblyName>Forte.EpiEasyEvents</AssemblyName>
+    <RootNamespace>Forte.EpiEasyEvents</RootNamespace>
+    <TargetFramework>net5.0</TargetFramework>
+  </PropertyGroup>
+  <ItemGroup>
+    <PackageReference Include="EPiServer.CMS" Version="12.3.0" />
+    <PackageReference Include="EPiServer.Framework" Version="12.3.0" />
+  </ItemGroup>
 </Project>
